StorySign_Platform/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
opencv-python>=4.8.0
# mediapipe>=0.10.0  # Not available for Python 3.13 yet
python-multipart==0.0.6
pydantic>=2.11.0
numpy>=1.24.0
pillow==10.1.0
python-json-logger==2.0.7
pyyaml==6.0.1
psutil>=5.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
# Database dependencies for TiDB support
sqlalchemy[asyncio]>=2.0.0
asyncmy>=0.2.9
pymysql>=1.1.0
# Authentication dependencies
bcrypt>=4.0.0
pyjwt>=2.8.0